Removed inconsistencies from the code

+++ src/main/java/com/thevoxelbox/voxelsniper/brush/SplatterVoxelBrush.java
@@ -206,7 +206,7 @@ public class SplatterVoxelBrush extends PerformerBrush {
 
                 if (temp >= GROW_PERCENT_MIN && temp <= GROW_PERCENT_MAX) {
                     v.sendMessage(ChatColor.AQUA + "Growth percent set to: " + String.format("%.2f", (double) temp / 100) + "%");
-                    this.growPercent = (int) temp;
+                    this.growPercent = temp;
                 } else {
                     v.sendMessage(ChatColor.RED + "Growth percent must be a decimal between 0.01 - 99.99!");
                 }
